LayoutTests/fast/text/text-orientation-parse-competition.html

<!DOCTYPE html>
<html>
	<head>
	<meta charset="utf-8">
	<script src="../../resources/js-test-pre.js"></script>
	</head>
	<body>
		<div id="test1" style="-webkit-writing-mode: vertical-rl; -webkit-text-orientation: sideways; text-orientation: upright">你好ABC</div>
		<div id="test2" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways; -webkit-text-orientation: upright">你好ABC</div>
		<script>
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).textOrientation", "upright");
		</script>
		<script src="../../resources/js-test-post.js"></script>
	</body>
</html>
